feat: Add SentenceTransformer class for vector embedding
This commit adds a new file, SentenceTransformer.py, which contains the implementation of the STBert class. This class is responsible for encoding text using the SentenceTransformer model. It also adds a new file, VectorEmbedding.py, which contains the VectorEmbeddingProvider class. This class uses the STBert class to provide vector embedding functionality.